Move Extensions outside service
