+++ content_store/lib/content_store.dart
@@ -9,7 +9,7 @@ export 'online_service/builders/get_preview_uri.dart';
 export 'db_service/builders/get_store_updater.dart';
 export 'db_service/builders/w3_get_collection.dart';
 export 'db_service/builders/w3_get_media.dart';
-export 'db_service/extensions/ext_cl_medias.dart';
+export 'extensions/ext_cl_medias.dart';
 export 'db_service/widgets/broken_image.dart';
 export 'db_service/widgets/shimmer.dart';
 export 'db_service/models/store_updater.dart';
