+++ content_store/lib/db_service/models/store_updter_ext_store.dart
@@ -15,8 +15,8 @@ import 'package:mime/mime.dart';
 import 'package:path/path.dart' as p;
 import 'package:store/store.dart';
 
-import '../extensions/ext_cldirectories.dart';
-import '../extensions/list_ext.dart';
+import '../../extensions/ext_cldirectories.dart';
+import '../../extensions/list_ext.dart';
 import '../../storage_service/models/file_system/models/cl_directories.dart';
 
 import 'share_files.dart';
