+++ content_store/lib/extensions/ext_cldirectories.dart
@@ -1,7 +1,7 @@
 import 'package:path/path.dart' as p;
 import 'package:store/store.dart';
 
-import '../../storage_service/models/file_system/models/cl_directories.dart';
+import '../storage_service/models/file_system/models/cl_directories.dart';
 import 'ext_cl_media.dart';
 
 extension PathExt on CLDirectories {
