+++ content_store/lib/online_service/providers/scanner.dart
@@ -5,7 +5,7 @@ import 'package:connectivity_plus/connectivity_plus.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:nsd/nsd.dart';
 
-import '../../db_service/extensions/list_ext.dart';
+import '../../extensions/list_ext.dart';
 import '../models/cl_server.dart';
 import '../models/network_scanner.dart';
